+++ Changelog.md
@@ -1,4 +1,9 @@
 # Changelog
+## [1.0.1] - 2026-05-22
+### Update
+- Corrected export size
+- Improved gaurd points logic
+
 ## [1.0.0] - 2026-03-18
 ### Added
 - Initial release